+++ packages/webpack-plugin-hash-output/sewing-kit.config.ts
@@ -1,8 +1,0 @@
-import {createPackage, Runtime} from '@sewing-kit/core';
-
-import {createSewingKitPackagePlugin} from '../../config/sewing-kit';
-
-export default createPackage((pkg) => {
-  pkg.runtime(Runtime.Node);
-  pkg.use(createSewingKitPackagePlugin());
-});